fix typo and rename folder to multiclass_extension(s)
