+++ qiskit_aqua/algorithms/components/multiclass_extensions/estimator.py
